+++ docs/source/conf.py
@@ -10,7 +10,7 @@ copyright = '2026, Ahmed Fawzy Gad'
 author = 'Ahmed Fawzy Gad'
 
 # The full version, including alpha/beta/rc tags.
-release = '3.6.0'
+release = '3.7.0'
 
 master_doc = 'index'