+++ pygad/nn/__init__.py
@@ -1,4 +1,4 @@
 from .nn import *
 
-__version__ = "1.2.2"
+__version__ = "1.2.3"